requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-oauth-toolkit==2.2.0
mysqlclient==2.1.1
Pillow==9.3.0
django-ses==3.2.2
django-cors-headers==3.13.0
python-dotenv==0.21.0
django-celery-beat==2.4.0
gunicorn==20.1.0